Added exports to all classes.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@884 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ SoundWrapper/src/OggStream.cpp
@@ -66,7 +66,7 @@ size_t OggStream::read_cb(void *ptr, size_t size, size_t nmemb, void *datasource
 
 int OggStream::seek_cb(void *datasource, ogg_int64_t offset, int whence)
 {
-	return fseek((FILE*)datasource, offset, whence);
+	return fseek((FILE*)datasource, (long)offset, whence);
 }
 
 int OggStream::close_cb(void *datasource)
